% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{get_boundary}
\alias{get_boundary}
\title{get_boundary}
\usage{
get_boundary(region)
}
\arguments{
\item{region}{region name stored in a vector 
(vector length depends on the granularity level of the region,
from national level down to county level, 
i.e. c(“USA”), c(“USA”,“Pennsylvania”), c(“USA
”,“Pennsylvania”, “Allegheny”)), and}
}
\value{
a spatial polygon outlining the region
}
\description{
get_boundary draws the geographical boundary of a specified region from GADM.
It takes in a vector of region name and returns a spatial polygon outlining the region.
}
\examples{
# Plotting the region of Pennsylvania
PA <- get_boundary(c("USA", "Pennsylvania"))
plot(PA)

# Plotting the environment conditions for Pennsylvania
PA_environment_6 <- mask(crop(environment_data[[6]], PA), PA)
plot(PA_environment_6)

}
