% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{get_environment_layers}
\alias{get_environment_layers}
\title{get_environment_layers}
\usage{
get_environment_layers(resolution = 10)
}
\arguments{
\item{resolution}{the spatial resolution of environmental layers: 10, 5, 2.5}
}
\value{
a list of environment raster stack corresponding to environment conditions in each month
}
\description{
get_environment_layers
}
\examples{
get_environment_layers()
}
