% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{get_maxent_predict}
\alias{get_maxent_predict}
\title{get_maxent_predict}
\usage{
get_maxent_predict(month, species_presence, climate, species_name)
}
\arguments{
\item{month}{an integer representing the month of the year (from 1 to 12)}

\item{species_presence}{a dataframe containing longitudes and latitudes of species observation}

\item{climate}{a raster stack of environmental layers in the corresponding month}

\item{species_name}{a length-2 vector containing genus and species name}
}
\value{
raw: raw probability raster

logis: logistic probability raster
}
\description{
get_maxent_predict
}
\examples{
# Example: predict mosquito distribution in June
month <- 6
mosquito_month <- mosquito_presence[mosquito_presence$month==month, c("lon", "lat")]
mosquito_month <- subset(mosquito_month, !is.na(lon) & !is.na(lat))
mosquito_predictions_6 <- get_maxent_predict(month, mosquito_month, environment_data[[6]], c("aedes", "aegypti"))
par(mfrow=c(1,2))
plot(mosquito_predictions_6$raw)
title("Aedes Aegypti Distribution Raw Predictions")
plot(mosquito_predictions_6$logis)
title("Aedes Aegypti Distribution Logistic Predictions")

}
