% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{plot_results}
\alias{plot_results}
\title{plot_results}
\usage{
plot_results(results)
}
\arguments{
\item{results}{a list of returned values from run_simulations}
}
\value{
None
}
\description{
plot_results plots the logistic probabilities 
and population sample points from multiple species simulations 
using a common color scale.
}
\examples{
plot_results(list(result1, result2))

}
