% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{run_simulation}
\alias{run_simulation}
\title{run_simulation}
\usage{
run_simulation(species_name, month, N, region, resolution = 10)
}
\value{
log

sample

range

N
}
\description{
run_simulation
}
\examples{
region1 <- c("USA", "Pennsylvania")
species_name1 <- c("Ixodes", "ricinus", "Linnaeus")
species_name2 <- c("mus", "musculus")
result1 <- run_simulation(species_name1, month, N, region1, resolution)
result2 <- run_simulation(species_name2, month, N, region1, resolution)

}
