% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{get_boundary}
\alias{get_boundary}
\title{get_boundary}
\usage{
get_boundary(region)
}
\arguments{
\item{region}{region name stored in a vector (vector length depends on the granularity level of the region)}
}
\value{
a spatial polygon outlining the region
}
\description{
get_boundary
}
\examples{
# Plotting the region of Pennsylvania
PA <- get_boundary(c("USA", "Pennsylvania"))
plot(PA)

# Plotting the environment conditions for Pennsylvania
PA_environment_6 <- mask(crop(environment_data[[6]], PA), PA)
plot(PA_environment_6)

}
