% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{get_boundary}
\alias{get_boundary}
\title{get_boundary}
\usage{
get_boundary(region)
}
\arguments{
\item{region}{region name stored in a vector (vector length depends on the granularity level of the region)}
}
\value{
a spatial polygon outlining the region
}
\description{
get_boundary
}
\examples{
get_boundary(c("USA")) 
get_boundary(c("USA","Pennsylvania")) 
get_boundary(c("USA","Pennsylvania", "Allegheny"))
}
