% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{get_maxent_predict}
\alias{get_maxent_predict}
\title{get_maxent_predict}
\usage{
get_maxent_predict(month, species_presence, climate, species_name)
}
\arguments{
\item{month}{an integer representing the month of the year (from 1 to 12)}

\item{species_presence}{a dataframe containing longitudes and latitudes of species observation}

\item{climate}{a raster stack of environmental layers in the corresponding month}

\item{species_name}{a length-2 vector containing genus and species name}
}
\value{
raw raw probability raster

logis logistic probability raster
}
\description{
get_maxent_predict
}
\examples{
get_maxent_predict(6, mosquito_presence, climate_jun, c("aedes", "aegypti"))
}
