% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{get_species_data}
\alias{get_species_data}
\title{get_species_data}
\usage{
get_species_data(species_name)
}
\arguments{
\item{species_name}{a length-2 vector containing genus and species name}
}
\value{
a dataframe of presence data for the specified species
}
\description{
get_species_data
}
\examples{
get_species_data(c("aedes", "aegypti"))
# example: plot mosquito species presence data
mosquito_presence <- get_species_data(c("aedes", "aegypti"))
mosquito_location <- mosquito_presence[, c("lon", "lat")]
# install.packages("maptools")
library(maptools)
data(wrld_simpl)
plot(wrld_simpl)
points(mosquito_location, col='blue', cex=0.5)
title("Aedes Aegypti Observations")

}
