% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{population_sampling}
\alias{population_sampling}
\title{population_sampling}
\usage{
population_sampling(maxent_predict_raw, maxent_predict_logis,
  region = c("ITA"), species, N, month_logis_base, month, graph_logis = F,
  graph_raw = F)
}
\arguments{
\item{maxent_predict_raw}{character containing a string to
print}
}
\value{
month_samples: sample points drawn from the given month

month_region_log: logistic probabilities in the given region in given month

month_N: the number of samples drawn from the given month proportional to the sample size from the reference month
}
\description{
population_sampling
}
\examples{
# Example: draw a sample of 2000 mosquitos from Pennsylvania
sampling_results <- population_sampling(mosquito_predictions_6$raw,
mosquito_predictions_6$logis,
region=c("USA", "Pennsylvania"),
species=c("aedes", "aegypti"), 
N=2000,
month_logis_base=mosquito_predictions_6$logis, 
month=6, graph_logis=T, graph_raw=T) 
plot(sampling_results$month_region_log)
points(sampling_results$month_samples, cex=0.5, pch=16, col="blue")
title("Logistic Probabilities and Aedes Aegypti Population Sampling for PA")
}
